requirements.txt
easydict==1.11
h5py==3.10.0
numpy==1.26.2
PyYAML==6.0
PyYAML==6.0.1
torch==2.1.0
